Feat: 투표 완료 모달 내에서 득표수 카운트, 카메라 이동 로직 처리, 단계 완료는 미구현
