+++ src/components/Modals/VoteModal.tsx
@@ -1,12 +1,16 @@
-import { useStorage, useMutation } from "~/liveblocks.config";
+import {
+  useStorage,
+  useMutation,
+  useOthers,
+  useBroadcastEvent,
+} from "~/liveblocks.config";
 import useModalStore from "@/store/useModalStore";
 import { useState } from "react";
-import { TemplateType } from "@/lib/types";
-import { useOthers, useSelf } from "~/liveblocks.config";
-import { ThirdStepProbTemplate } from "@/lib/types";
+import { ThirdStepProbTemplate, TemplateType } from "@/lib/types";
 
 type voteCandidate = 1 | 2 | 3 | 4 | 5;
 const VoteModal = () => {
+  const broadcast = useBroadcastEvent();
   const { setModalState, setModalType } = useModalStore();
   const [vote, setVote] = useState<1 | 2 | 3 | 4 | 5 | null>(null);
   const [voteCompleted, setVoteCompleted] = useState<boolean>(false);
@@ -53,29 +57,32 @@ const VoteModal = () => {
 
       if (newTotalCount === totalPeople) {
         //투표 알고리즘
-        const voteCountObject = voteCount.toObject();
-        let maxVotes = 0;
-        let winningVote = null;
+        // const voteCountObject = voteCount.toObject();
+        // let maxVotes = 0;
+        // let winningVote = null;
 
-        for (const [key, value] of Object.entries(voteCountObject)) {
-          console.log(key, value);
-          if (value > maxVotes) {
-            maxVotes = value;
-            winningVote = key;
-          }
-        }
+        // for (const [key, value] of Object.entries(voteCountObject)) {
+        //   console.log(key, value);
+        //   if (value > maxVotes) {
+        //     maxVotes = value;
+        //     winningVote = key;
+        //   }
+        // }
 
-        console.log("최다 득표 번호:", winningVote, "번", "득표수:", maxVotes);
+        // console.log("최다 득표 번호:", winningVote, "번", "득표수:", maxVotes);
 
-        //voteStore초기화. 테스트용.
-        voteCount.update({
-          "1": 0,
-          "2": 0,
-          "3": 0,
-          "4": 0,
-          "5": 0,
-        });
-        voteList.set("totalCount", 0);
+        // //voteStore초기화. 테스트용.
+        // voteCount.update({
+        //   "1": 0,
+        //   "2": 0,
+        //   "3": 0,
+        //   "4": 0,
+        //   "5": 0,
+        // });
+        // voteList.set("totalCount", 0);
+        broadcast({ type: "VOTE_END", message: "sync Complete!" });
+        setModalType("voteComplete");
+        setModalState(true);
       }
     },
     [voteList],
